attempt to fix pipeline
